requirements.txt
openpyxl==3.1.2 
selenium==4.17.2
requests==2.31.0
PyYAML==6.0.1
webdriver-manager==4.0.1
python-dotenv